Added distance validation for the ride and implemented a BaseViewModel

+++ app/src/main/java/net/rafgpereira/transpoapp/data/network/IApiService.kt
@@ -23,7 +23,7 @@ interface IApiService {
     @GET("/ride/{userId}?driver_id={driverId}")
     suspend fun getHistory(
         @Path("userId") userId: String,
-        @Path("driverId") driverId: Long
+        @Path("driverId") driverId: Int
     ) : Response<HistoryResult>
 
     companion object {
